backend/requirements.txt
django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.3
httpx==0.27.0
django-celery-beat==2.6.0
